Conflicts:
	ems/src/main/java/com/vts/ems/master/controller/MasterController.java
	ems/src/main/webapp/view/masters/Circular.jsp

+++ ems/src/main/java/com/vts/ems/master/dao/MasterDaoImpl.java
@@ -846,7 +846,7 @@ private static final String DUPLICATEDOCQUALIFICATION = "SELECT COUNT(docqualifi
 			Query query = manager.createNativeQuery("SELECT MAX(circularid)  FROM chss_circular_list");	
 			
 			BigInteger result = (BigInteger) query.getSingleResult();
-			return result.intValue();
+			return result.longValue();
 		} catch (Exception e) {
 			e.printStackTrace();
 			return 0;
